git-svn-id: svn://svn.jdownloader.org/jdownloader/trunk@941 ebf7c1c2-ba36-0410-9fe8-c592906822b4

+++ src/jd/captcha/specials/UploadedTo.java
@@ -69,7 +69,7 @@ public class UploadedTo {
             // l.reduceWhiteNoise(2);
             // l.toBlackAndWhite(0.6);
 
-            ret[i] = l.getSimplified(captcha.owner.getJas().getInteger("simplifyFaktor"));
+            ret[i] = l.getSimplified(captcha.owner.getJas().getDouble("simplifyFaktor"));
 
         }
         return ret;
